Add scoreboard api (Closes #1)

+++ src/main/java/ai/codemap/codemap/form/JudgeToMain.java
@@ -8,11 +8,12 @@ import lombok.Setter;
 @Getter
 @Setter
 public class JudgeToMain {
+
     private Long submissionId;
     private String status;
-    private double time;
+    private Double time;
     private Long memory;
     private String compilerMessage;
     private String output;
-    private int score;
+    private Long score;
 }
